// src/app/services/application/actions.enum.ts
export enum ApplicationActions {
  INITIALIZING = 'initializing',
  STATECHANGE = 'statechange',
  SETSTATE = 'setstate',
  SETSTATETOROUTE = 'setstatetoroute'
}

export enum AppStateChildrens {
  APPLICATIONPARAMS = 'applicationParams',
  DEVICE = 'device',
  STYLEGUIDE = 'styleguide',
  ROUTES = 'routes',
  DEFAULTDATA = 'defaultData',
  ROUTERDATA = 'routerData'
}